+++ src/pages/EstateManage/EstateBoard/js/EstateBoard.js
@@ -1,0 +1,13 @@
+import { useState } from "react";
+import axios from "axios";
+
+function EstateBoard() {
+
+  return (
+    <div className="container">
+
+    </div>
+  );
+}
+
+export default EstateBoard;
